+++ src/main/java/com/monorama/iot_server/interceptor/UserIdInterceptor.java
@@ -1,4 +1,4 @@
-package com.monorama.iot_server.intercepter;
+package com.monorama.iot_server.interceptor;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
